Take modern mode of Vuex
