+++ components/LocaleSwitcher.vue
@@ -32,7 +32,7 @@
         d="M11.25,8.4 L12.87,8.4 L15.89,16 L14.27,16 L13.55,14 L10.49,14 L9.76,16 L8.23,16 L11.25,8.4 Z M13.13,12.82 L12.01,9.76 L10.9,12.82 L13.13,12.82 Z"
       />
     </g>
-  </svg> 
+  </svg>
 </template>
 
 <script lang="ts">
