+++ pages/about.vue
@@ -84,7 +84,7 @@
   import PageBase from "~/lib/page-base";
   import HaiiroButton from "~/components/HaiiroButton.vue";
 
-  import { name as PixelsNamespace } from "~/store/modules/pixels";
+  import { name as PixelsNamespace } from "~/store/pixels";
   const PixelsStore = namespace(PixelsNamespace);
 
   import enData from "~/contents/en/about.md";
