+++ pages/contact.vue
@@ -68,7 +68,7 @@
   import { Component } from "nuxt-property-decorator";
   import { namespace } from "vuex-class";
 
-  import { name as PixelsNamespace } from "~/store/modules/pixels";
+  import { name as PixelsNamespace } from "~/store/pixels";
   const PixelsStore = namespace(PixelsNamespace);
 
   import PageBase from "~/lib/page-base";
